made map generated c files to include custom typedef file to make more portable

+++ algorithm-testing/dijkstra.h
@@ -5,6 +5,6 @@
 
 #define MAX_DISTANCE 1000000
 #define MAX_PATH_LENGTH 20
-extern uint8_t path[];
+extern int path[];
 void dijkstra(int, int);
 #endif
